Redesign of tile model, new flow
Instead of calculating the mesh with no prospects of sane debuging, now
create models in blender and iron out the import problems.

Blender took care of UV mapping, so it looks OK (images and other
resources added to source tree); and transform data was generally
imported alright except for a minor fix to rotation.

Now, the TileBehaviour does some minor import fixes on Start(), and then
manipulates the given mesh on Update() if need be. Not much is assumed
about the order of the vertices in the imported mesh or the duplicity of
vertices
...BUT...
strong assumptions are made about the DISTINCT vertices of the mesh:
there should be exactly 10 or 12 of them (pents/hexes) and their vector3
values are used for logic.

Signed-off-by: dorimedini <dorimedini@gmail.com>