Why did we delete all those things now? Well...
... after making the gravity work, with no visible changes in the
working directory, I found out that .meta files should actually be
source controlled... damn. So, instead of adding eleventy billion mostly
useless files, I removed unused packages and brought that number down
with a higher percentage of non-useless files.

The bake-Y-position fix that solved the gravity issue is buried
somewhere in the .meta changes to the locomotion and idle animations.

Signed-off-by: dorimedini <dorimedini@gmail.com>