Implemented projectile fire & player launch
Works like so:
- ShootingCharacters can fire projectiles
- Firing a projectile invokes an RPC call which creates a local instance
  of the projectile on every machine
- All colliders are active
This isn't so good, I should do Photon.Instantiate maybe instead...? But
force needs to be added on instantiate!

The projectile, when colliding with a tile, invokes the extension logic
via RPC call (special class designed for pillar extensions). This call
works well, as it takes care to only launch *local* players in the air
(other players will see the movement updated anyway).

Signed-off-by: dorimedini <dorimedini@gmail.com>