Now inflict damage
The projectile that hits a player broadcasts a message to all players
the specific player (identified by user ID) was hit.

The hit player takes damage on all clients.

It may be better to make the player's PhotonView watch the Health
component but I'd say the less photon views the better...

Signed-off-by: dorimedini <dorimedini@gmail.com>