Better projectile syncing
I don't know exactly how this worked until now... maybe it was the cause
of the null-reference error when accessing the activeProjectiles dict...
but anyway, now projectiles only hve colliders if they were fired by the
local player. This is implemented by making the local player instaniate
a projectile WITH a collider, then broadcast an RPC to all other players
to instantiate a projectile WITHOUT a collider.

The local player will now get immediate response when the projectile
hits something.

Signed-off-by: dorimedini <dorimedini@gmail.com>