Some projectile bugfixes...
Previously, won direct hit from projectile, all players in the room get
the RPC. I wanted players to add explosive force to themselves only, so
I used photonView.Owner.UserId to compare to the target userID.

Problem is, ExplosionController doesn't have an owner (not
network-instantiated?).

To solve this, when initially spawning the local player in the
NetworkManager I set the user ID as a variable in the
ExplosionController for later access. Ths value should be constant
throughout the match anyway.

Signed-off-by: dorimedini <dorimedini@gmail.com>