Sphere radius now user-defined
The implications were: the GeoSphere had to be generated after the user
settings were loaded. NetworkManager now gives the pillar controller
and the spawn manager a reference to the geosphere after initializing.

Note: to set the radius, edit the radius value in the GUI panel, and
then restart the game.

Signed-off-by: dorimedini <dorimedini@gmail.com>