Grapple v2.0
We have a new model - chain (with individual links!) and a hook at the
end. No need to destroy and re-instantiate when the length / position of
the grapple shot changes, have a script for dynamically adding /
removing segments and turn individual links on/off when needed.

Removed old rope stuff, git thinks the meta files are similar so
whatever

Signed-off-by: dorimedini <dorimedini@gmail.com>