Explosion sync 'working'
Now, the exploding projectile sends individual RPCs to hit targets with
more robust information: the force vector to apply. This means the local
force applied will be identical to the force applied on the remote
machine, although not from the same source position (the player may have
travelled in the process).

This required bringing the movement controller up to speed with the fact
that original speed on the rigidbody may exceed the magnitude-1 movement
speed allowed normally (previous commit).

Signed-off-by: dorimedini <dorimedini@gmail.com>