Now have targeting sequence in place
Currently can press and hold fire button to initiate targeting. Targeting
continues so long as fre button is pressed and enemy player is targetable
(currently only check if in scope, not if something is in the way).

Also fixed TODOs in the places where targeting indication should be sent
across the network to the targeted player (among other TODOs sprinkled
with bich)

Signed-off-by: dorimedini <dorimedini@gmail.com>