All explosions now handled by explosion controller
Broadcasting an explosion now triggers graphics and hits on all players
in the game.

Use the friendlyFire parameter to decide if this explosion should damage
the shooter / explosion instigator.

Now that the shooter ID is also needed to identify the local offline
player, we also need to handle the case where the shooter ID is null, in
which case we replace it with an empty string.

Signed-off-by: dorimedini <dorimedini@gmail.com>